Game-Of-Life
The Game of Life is an archetypal checkerboard simulation. It was invented by the English mathematician
John H. Conway for exploring certain formal situations involving reproduction and growth (see Scientific
American, October 1970, p. 120). The game is a successor to the work of John von Neumann on cellular
automata.
A checkerboard simulation is a model wherein the world is broken into a rectangular array of cells. Each cell
may interact only with its neighboring cells, the interactions occurring once per simulated clock pulse. The
rules for a local change are well defined and simple, but predicting global change is impractical without computer
simulation. The Game of Life has the following rules.

GAME OF LIFE RULES
1. A cell is either empty, indicated by a blank, or alive, indicated by an X.
2. Each cell is the center of a 3 􀂉 3 square grid of cells, which contains its eight neighbors.
3. A cell that is empty at time t becomes alive at time t + 1 if and only if exactly three neighboring
cells were alive at time t.
4. A cell that is alive at time t remains alive at time t + 1 if and only if either two or three neighboring
cells were alive at time t. Otherwise, it dies for lack of company ( < 2 ) or overcrowding ( > 3 ).
5. The simulation is conducted, in principle, on an infinite two-dimensional grid.
A simulation starts with an initial configuration of Xs on the grid. The following example of a simple repeating
life form is called a blinker, as it repeats every two generations, or time steps.